IkarusIA/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
twilio==9.3.0
firebase-admin==6.5.0
